+++ src/Whim.Updater.Tests/Data.cs
